pyproject.toml
[project]
name = "sth"
authors = [
    { name = "DAS", email = "digitalasiasolusindo@gmail.com"}
]
description = "STH Module"
requires-python = ">=3.10"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    # "frappe~=15.0.0" # Installed and managed by bench.
    "pandas~=2.3.3",
    "reportlab~=4.4.9",
    "python-gnupg~=0.5.6",
    "paramiko~=4.0.0",
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

# These dependencies are only installed when developer mode is enabled
[tool.bench.dev-dependencies]
# package_name = "~=1.1.0"

[tool.bench.frappe-dependencies]
frappe = ">=15.0.0,<16.0.0"
erpnext = ">=15.0.0,<16.0.0"